Mejoras en el Codificador
Esta Actualizacion trae los siguientes cambios en el codificador y maestro de codificador

-Ahora en las tablas se muestra el nombre de los tipos de producto , lineas y sublineas.

-Se mejora la lectura del codigo, ya que se separa el texto javascript en un archivo independiente por cada vista.

- se mejora la visualizacion de las columnas, se elimina la columna index y se agrega la columna 'ultima Actualizacion' para dar un vistazo a la fecha y hora que fue modificado el registro por ultima vez

-bugfix
