+++ app/Http/Controllers/ProdCievCodCodigoController.php
@@ -10,6 +10,7 @@ use App\CodMedida;
 use App\CodSublinea;
 use App\CodTipoProducto;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Yajra\DataTables\DataTables;
 
 class ProdCievCodCodigoController extends Controller
@@ -18,9 +19,20 @@ class ProdCievCodCodigoController extends Controller
     {
         if ($request->ajax()) {
 
-            $data = CodCodigo::latest()->get();
+            $data = DB::table('cod_codigos')
+                ->leftJoin('cod_tipo_productos','cod_codigos.cod_tipo_producto_id','=','cod_tipo_productos.id')
+                ->leftJoin('cod_lineas','cod_codigos.cod_lineas_id','=','cod_lineas.id')
+                ->leftJoin('cod_sublineas','cod_codigos.cod_sublineas_id','=','cod_sublineas.id')
+                ->leftJoin('cod_medidas','cod_codigos.cod_medidas_id','=','cod_medidas.id')
+                ->leftJoin('cod_materials','cod_codigos.cod_materials_id','=','cod_materials.id')
+                ->leftJoin('cod_caracteristicas','cod_codigos.cod_caracteristicas_id','=','cod_caracteristicas.id')
+                ->select('cod_codigos.codigo as codigo','cod_codigos.coments as coment','cod_codigos.descripcion as desc','cod_codigos.usuario','cod_codigos.usuario_aprobo',
+                    'cod_codigos.arte','cod_codigos.estado','cod_codigos.area','cod_codigos.costo_base','cod_codigos.generico','cod_codigos.created_at',
+                    'cod_codigos.updated_at','cod_tipo_productos.name as tp','cod_lineas.name as lin','cod_sublineas.name as subl','cod_medidas.name as med','cod_materials.name as mat',
+                    'cod_caracteristicas.name as car','cod_codigos.id as id')
+                ->get();
+
             return DataTables::of($data)
-                ->addIndexColumn()
                 ->addColumn('Opciones', function($row){
                     $btn = '<a href="javascript:void(0)" data-toggle="tooltip"  data-id="'.$row->id.'" data-original-title="Editar" class="edit btn btn-primary btn-sm editCodigo" id="edit-btn">Editar</a>';
                     $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip"  data-id="'.$row->id.'" data-original-title="Eliminar" class="btn btn-danger btn-sm deleteCodigo">Eliminar</a>';
@@ -37,15 +49,14 @@ class ProdCievCodCodigoController extends Controller
         CodCodigo::updateOrCreate(['id'     => $request->medida_id],
             [   'codigo'                    => $request->codigo,
                 'coments'                   => $request->coments,
-                'cod_tipo_producto_id'      => $request->ctipoproducto_id,
-                'cod_lineas_id'             => $request->clineas_id,
+                'cod_tipo_producto_id'      => $request->tipoproducto_id,
+                'cod_lineas_id'             => $request->lineas_id,
                 'cod_sublineas_id'          => $request->sublineas_id,
                 'cod_medidas_id'            => $request->medida_id,
                 'cod_caracteristicas_id'    => $request->caracteristica_id,
                 'cod_materials_id'          => $request->material_id,
                 'descripcion'               => $request->descripcion
             ]);
-
         return response()->json(['success'=>'Medida Guardada Correctamente.']);
     }
 
@@ -132,7 +143,6 @@ class ProdCievCodCodigoController extends Controller
             }
             return response()->json($Array);
         }
-
     }
 
     public function lns (Request $request)
@@ -143,7 +153,6 @@ class ProdCievCodCodigoController extends Controller
                 $Array1[$v1->cod] = $v1->abreviatura;
             }
             return response()->json($Array1);
-
         }
     }
 
@@ -156,7 +165,6 @@ class ProdCievCodCodigoController extends Controller
                 $Array[$v1->cod] = $v1->abreviatura;
             }
             return response()->json($Array);
-
         }
     }
 
@@ -168,7 +176,6 @@ class ProdCievCodCodigoController extends Controller
                 $Array[$v1->cod] = $v1->abreviatura;
             }
             return response()->json($Array);
-
         }
     }
 
@@ -180,7 +187,6 @@ class ProdCievCodCodigoController extends Controller
                 $Array[$v1->cod] = $v1->abreviatura;
             }
             return response()->json($Array);
-
         }
     }
 
@@ -192,7 +198,6 @@ class ProdCievCodCodigoController extends Controller
                 $Array[$v1->id] = $v1->denominacion;
             }
             return response()->json($Array);
-
         }
     }
