+++ resources/views/ProductosCIEV/Maestros/caracteristicas_show.blade.php
@@ -1,10 +1,10 @@
 @extends('layouts.dashboard')
 
-@section('page_title', 'Maestros')
+@section('page_title', 'Maestros (Caracteristicas)')
 
 @section('module_title', 'Caracteristicas')
 
-@section('subtitle', 'Este modulo permite ver, crear y editar caracteristicas.')
+@section('subtitle', 'Es la máxima particularización de las lineas, con ellas se da un detalle mayor de las piezas.')
 
 @section('breadcrumbs')
     {{ Breadcrumbs::render('Prod_ciev_maestros_caracteristicas') }}
@@ -14,7 +14,7 @@
     @inject('Lineas','App\Services\Lineas')
     <div class="col-lg-4">
         <div class="form-group">
-            <a class="btn btn-primary" href="javascript:void(0)" id="CrearCaracteristica">Crear Nueva Caracteristica</a>
+            <a class="btn btn-primary" href="javascript:void(0)" id="CrearCaracteristica">Nuevo</a>
         </div>
     </div>
     <div class="row">
@@ -25,13 +25,13 @@
                         <table class="table table-striped first data-table">
                             <thead>
                                 <tr>
-                                    <th>#</th>
-                                    <th>Codigo de linea</th>
-                                    <th>Codigo de Sublinea</th>
+                                    <th>Linea</th>
+                                    <th>Sublinea</th>
                                     <th>Codigo</th>
-                                    <th>Nombre Caracteristica</th>
+                                    <th>Nombre</th>
                                     <th>Comentarios</th>
-                                    <th width="280px">Opciones</th>
+                                    <th>Ultima Actualizacion</th>
+                                    <th>Opciones</th>
                                 </tr>
                             </thead>
                             <tbody>
@@ -109,136 +109,7 @@
 
 
     @push('javascript')
-        <script type="text/javascript">
-            $(document).ready(function(){
-                $(function () {
-                    $.ajaxSetup({
-                        headers: {
-                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                        }
-                    });
-
-                    var table = $('.data-table').DataTable({
-                        processing: true,
-                        serverSide: true,
-                        ajax: "{{ route('ProdCievCodCaracteristica.index') }}",
-                        columns: [
-                            {data: 'DT_RowIndex', name: 'DT_RowIndex'},
-                            {data: 'car_lineas_id', name: 'car_lineas_id'},
-                            {data: 'car_sublineas_id', name: 'car_sublineas_id'},
-                            {data: 'cod', name: 'cod'},
-                            {data: 'name', name: 'name'},
-                            {data: 'coments', name: 'coments'},
-                            {data: 'Opciones', name: 'Opciones', orderable: false, searchable: false},
-                        ],
-                        language: {
-                            // traduccion de datatables
-                            processing: "Procesando...",
-                            search: "Buscar&nbsp;:",
-                            lengthMenu: "Mostrar _MENU_ registros",
-                            info: "Mostrando registros del _START_ al _END_ de un total de _TOTAL_ registros",
-                            infoEmpty: "Mostrando registros del 0 al 0 de un total de 0 registros",
-                            infoFiltered: "(filtrado de un total de _MAX_ registros)",
-                            infoPostFix: "",
-                            loadingRecords: "Cargando...",
-                            zeroRecords: "No se encontraron resultados",
-                            emptyTable: "Ningún registro disponible en esta tabla :C",
-                            paginate: {
-                                first: "Primero",
-                                previous: "Anterior",
-                                next: "Siguiente",
-                                last: "Ultimo"
-                            },
-                            aria: {
-                                sortAscending: ": Activar para ordenar la columna de manera ascendente",
-                                sortDescending: ": Activar para ordenar la columna de manera descendente"
-                            }
-                        }
-                    });
-
-                    $('#CrearCaracteristica').click(function () {
-                        $('#saveBtn').val("create-caracteristica");
-                        $('#caracteristica_id').val('');
-                        $('#caracteristicaForm').trigger("reset");
-                        $('#modelHeading').html("Crear Nueva Caracteristica");
-                        $('#caracteristicamodal').modal('show');
-                    });
-
-                    $('body').on('click', '.editcaracteristica', function () {
-
-                        var caracteristica_id = $(this).data('id');
-                        $.get("{{ route('ProdCievCodCaracteristica.index') }}" +'/' + caracteristica_id +'/edit', function (data) {
-                            $('#modelHeading').html("Editar Caracteristica");
-                            $('#saveBtn').val("edit-caracteristica");
-                            $('#caracteristicamodal').modal('show');
-                            $('#caracteristica_id').val(data.id);
-                            $('#cod').val(data.cod);
-                            $('#car_lineas_id').val(data.car_lineas_id);
-                            $('#car_sublineas_id').val(data.car_sublineas_id);
-                            $('#name').val(data.name);
-                            $('#abreviatura').val(data.abreviatura);
-                            $('#coments').val(data.coments);
-                        })
-                    });
-
-
-                    $('#saveBtn').click(function (e) {
-                        e.preventDefault();
-                        //$(this).html('Guardando...');
-                        $.ajax({
-                            data: $('#caracteristicaForm').serialize(),
-                            url: "{{ route('ProdCievCodCaracteristica.store') }}",
-                            type: "POST",
-                            dataType: 'json',
-                            success: function (data) {
-
-                                $('#caracteristicaForm').trigger("reset");
-                                $('#caracteristicamodal').modal('hide');
-                                table.draw();
-                                toastr.success("Registro Guardado con Exito!");
-                                //   $(this).html('Crear');
-                            },
-                            error: function (data) {
-                                console.log('Error:', data);
-                                $('#saveBtn').html('Guardar Cambios');
-                            }
-                        });
-                    });
-
-                    $('body').on('click', '.deletecaracteristica', function () {
-
-                        var caracteristica_id = $(this).data("id");
-                        if(confirm("¿Esta seguro de Eliminar?")) {
-                            $.ajax({
-                                type: "DELETE",
-                                url: "{{ route('ProdCievCodCaracteristica.store') }}" + '/' + caracteristica_id,
-                                success: function (data) {
-                                    table.draw();
-                                    toastr.success("Registro eliminado con exito");
-                                },
-                                error: function (data) {
-                                    console.log('Error:', data);
-                                    toastr.danger("Error al eliminar el registro");
-                                }
-                            });
-                        }
-                    });
-                });
-
-                $('#car_lineas_id').on('change', function () {
-                    var lineas_id = $(this).val();
-                    if ($.trim(lineas_id) != ''){
-                        $.get('getsublineas',{lineas_id: lineas_id}, function(getsublineas) {
-                            $('#car_sublineas_id').empty();
-                            $('#car_sublineas_id').append("<option value=''>Seleccionar una sublinea</option>");
-                            $.each(getsublineas, function (index, value) {
-                                $('#car_sublineas_id').append("<option value='" + index + "'>"+ value +"</option>");
-                            })
-                        });
-                    }
-                });
-            });
-        </script>
+        <script type="text/javascript" src="/JsGlobal/Codificador/Maestros/Caracteristicas.js"></script>
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.0/jquery.validate.js"></script>
         <script src="https://cdn.datatables.net/1.10.16/js/jquery.dataTables.min.js"></script>
