+++ resources/views/ProductosCIEV/Maestros/sublineas_show.blade.php
@@ -1,20 +1,20 @@
 @extends('layouts.dashboard')
 
-@section('page_title', 'Maestros')
+@section('page_title', 'Maestros (Sublineas)')
 
 @section('module_title', 'Sublineas')
 
-@section('subtitle', 'sbutitulo.')
+@section('subtitle', 'División de las lineas que permiten granularizar las características de las piezas.')
 
 @section('breadcrumbs')
     {{ Breadcrumbs::render('Prod_ciev_maestros_sublineas') }}
 @stop
 
 @section('content')
-    @inject('Lineas','App\Services\Lineas')
+    @inject('TipoProductos','App\Services\TipoProductos')
     <div class="col-lg-4">
         <div class="form-group">
-            <a class="btn btn-primary" href="javascript:void(0)" id="CrearSubLineas">Crear Nueva Sublinea</a>
+            <a class="btn btn-primary" href="javascript:void(0)" id="CrearSubLineas">Nuevo</a>
         </div>
     </div>
     <div class="row">
@@ -25,12 +25,13 @@
                         <table class="table table-striped first data-table">
                             <thead>
                                 <tr>
-                                    <th>#</th>
-                                    <th>Codigo de linea</th>
+                                    <th>Tipo de Producto</th>
+                                    <th>Linea</th>
                                     <th>Codigo</th>
                                     <th>Nombre Sublinea</th>
                                     <th>Comentarios</th>
-                                    <th width="280px">Opciones</th>
+                                    <th>Ultima Actualizacion</th>
+                                    <th>Opciones</th>
                                 </tr>
                             </thead>
                             <tbody>
@@ -55,15 +56,21 @@
                     <form id="sublineaForm" name="sublineaForm" class="form-horizontal">
                         <input type="hidden" name="sublinea_id" id="sublinea_id">
                         <div class="form-group">
-                            <label for="name" class="col-sm-6 control-label">Linea:</label>
+                            <label for="name" class="col-sm-6 control-label">Tipo Producto:</label>
                             <div class="col-sm-12">
-                                <select class="custom-select" name="lineas_id" id="lineas_id">
-                                    @foreach ( $Lineas->get() as $index => $Linea)
-                                        <option value="{{ $index }}" {{ old('lineas_id') == $index ? 'selected' : ''}}>
-                                            {{ $Linea }}
+                                <select class="custom-select" name="tipoproductos_id" id="tipoproductos_id">
+                                    @foreach( $TipoProductos->get() as $index => $TipoProducto)
+                                        <option value="{{ $index }}" {{ old('tipoproductos_id') == $index ? 'selected' : ''}}>
+                                            {{ $TipoProducto }}
                                         </option>
                                     @endforeach
                                 </select>
+                            </div>
+                        </div>
+                        <div class="form-group">
+                            <label for="name" class="col-sm-6 control-label">Linea:</label>
+                            <div class="col-sm-12">
+                                <select class="custom-select" name="lineas_id" id="lineas_id"></select>
                             </div>
                         </div>
                         <div class="form-group">
@@ -100,130 +107,12 @@
         </div>
     </div>
 
-
     @push('javascript')
-        <script type="text/javascript">
-            $(function () {
-
-                $.ajaxSetup({
-                    headers: {
-                        'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                    }
-                });
-
-                var table = $('.data-table').DataTable({
-                    processing: true,
-                    serverSide: true,
-                    ajax: "{{ route('ProdCievCodSublinea.index') }}",
-                    columns: [
-                        {data: 'DT_RowIndex', name: 'DT_RowIndex'},
-                        {data: 'lineas_id', name: 'lineas_id'},
-						{data: 'cod', name: 'cod'},
-                        {data: 'name', name: 'name'},
-                        {data: 'coments', name: 'coments'},
-                        {data: 'Opciones', name: 'Opciones', orderable: false, searchable: false},
-                    ],
-                    language: {
-                        // traduccion de datatables
-                        processing: "Procesando...",
-                        search: "Buscar&nbsp;:",
-                        lengthMenu: "Mostrar _MENU_ registros",
-                        info: "Mostrando registros del _START_ al _END_ de un total de _TOTAL_ registros",
-                        infoEmpty: "Mostrando registros del 0 al 0 de un total de 0 registros",
-                        infoFiltered: "(filtrado de un total de _MAX_ registros)",
-                        infoPostFix: "",
-                        loadingRecords: "Cargando...",
-                        zeroRecords: "No se encontraron resultados",
-                        emptyTable: "Ningún registro disponible en esta tabla :C",
-                        paginate: {
-                            first: "Primero",
-                            previous: "Anterior",
-                            next: "Siguiente",
-                            last: "Ultimo"
-                        },
-                        aria: {
-                            sortAscending: ": Activar para ordenar la columna de manera ascendente",
-                            sortDescending: ": Activar para ordenar la columna de manera descendente"
-                        }
-                    }
-                });
-
-                $('#CrearSubLineas').click(function () {
-                    $('#saveBtn').val("create-sublinea");
-                    $('#sublinea_id').val('');
-                    $('#sublineaForm').trigger("reset");
-                    $('#modelHeading').html("Crear Nueva Sublinea");
-                    $('#sublineamodal').modal('show');
-                });
-
-                $('body').on('click', '.editsublinea', function () {
-
-                    var sublinea_id = $(this).data('id');
-                    $.get("{{ route('ProdCievCodSublinea.index') }}" +'/' + sublinea_id +'/edit', function (data) {
-                        $('#modelHeading').html("Editar Sublinea");
-                        $('#saveBtn').val("edit-sublinea");
-                        $('#sublineamodal').modal('show');
-                        $('#sublinea_id').val(data.id);
-                        $('#cod').val(data.cod);
-                        $('#lineas_id').val(data.lineas_id);
-                        $('#name').val(data.name);
-                        $('#abreviatura').val(data.abreviatura);
-                        $('#coments').val(data.coments);
-                    })
-                });
-
-
-                $('#saveBtn').click(function (e) {
-                    e.preventDefault();
-                    //$(this).html('Guardando...');
-                    $.ajax({
-                        data: $('#sublineaForm').serialize(),
-                        url: "{{ route('ProdCievCodSublinea.store') }}",
-                        type: "POST",
-                        dataType: 'json',
-                        success: function (data) {
-
-                            $('#sublineaForm').trigger("reset");
-                            $('#sublineamodal').modal('hide');
-                            table.draw();
-                            toastr.success("Registro Guardado con Exito!");
-                            //   $(this).html('Crear');
-                        },
-                        error: function (data) {
-                            console.log('Error:', data);
-                            $('#saveBtn').html('Guardar Cambios');
-                        }
-                    });
-                });
-
-                $('body').on('click', '.deletesubLinea', function () {
-
-                    var sublinea_id = $(this).data("id");
-                    if(confirm("¿Esta seguro de Eliminar?")) {
-                        $.ajax({
-                            type: "DELETE",
-                            url: "{{ route('ProdCievCodSublinea.store') }}" + '/' + sublinea_id,
-                            success: function (data) {
-                                table.draw();
-                                toastr.success("Registro Eliminado con exito");
-                            },
-                            error: function (data) {
-                                console.log('Error:', data);
-                                toastr.danger("Error al eliminar el registro");
-                            }
-                        });
-                    }
-                });
-            });
-        </script>
-        {{--<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.min.css" />--}}
-        {{--<link href="https://cdn.datatables.net/1.10.16/css/jquery.dataTables.min.css" rel="stylesheet">--}}
-        {{--<link href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css" rel="stylesheet">--}}
+        <script type="text/javascript" src="/JsGlobal/Codificador/Maestros/Sublineas.js"></script>
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-validate/1.19.0/jquery.validate.js"></script>
         <script src="https://cdn.datatables.net/1.10.16/js/jquery.dataTables.min.js"></script>
         <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
         <script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap4.min.js"></script>
-
     @endpush
 @endsection
